aegis_core/requirements.txt
# API & Server
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0

# AI & Orchestration
langchain>=0.1.13
langchain-core>=0.1.33
langchain-google-genai>=1.0.1
langgraph>=0.0.26

# Model Context Protocol (FastMCP)
mcp>=1.0.0

# Dashboard & Networking
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.1

langchain-mcp-adapters

neo4j
starlette
httpx
apscheduler